moved fakedata generation to separate module
